requirements.txt
# Core FastAPI dependencies
fastapi[all]==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Environment and configuration
python-dotenv==1.0.0

# API clients
requests==2.31.0
assemblyai==0.21.0
google-generativeai==0.3.2

# Additional utilities
pathlib-extensions==0.1.0

# Development dependencies (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
